raw-schema/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
google-cloud-bigquery